Fix airbrake gem to version 5.3.0
We have to fix to 5.3 because of a breaking change in 5.4. 5.4 updates its dependency on [airbrake-ruby](https://github.com/airbrake/airbrake/blob/master/CHANGELOG.md#v540-june-6-2016) to 1.4, and it in turn now validates that both `AIRBRAKE_PROJECT_KEY` and `AIRBRAKE_PROJECT_ID` exist, and that `AIRBRAKE_PROJECT_KEY` is an integer.

https://github.com/airbrake/airbrake-ruby/pull/87

Its not catastrophic; we simply need to add `AIRBRAKE_PROJECT_KEY=1` to our environments and once there we can then update this gem past version 5.3.

However we're putting that task off till later, and fixing the version for now.