Fix issue with Airbrake gem and project_id
It looks like the info in b3b39a0 was incorrect. It is true that we now need to set `AIRBRAKE_PROJECT_ID` and that the value must be an integer if we are using a later version of Airbrake. What was incorrect was the fact that the value comes from an environment variable.

In fact we typically set it in an initializer. Hence this change removes the block on airbrake being no greater than version 5.3.0.